+++ src/bot/requirements.txt
@@ -1,82 +1,8 @@
-asn1crypto==0.24.0
-atomicwrites==1.3.0
-attrs==19.3.0
-backports-abc==0.5
-backports.shutil-get-terminal-size==1.0.0
-bleach==3.1.1
-certifi==2019.11.28
+certifi==2020.4.5.1
 chardet==3.0.4
-configparser==4.0.2
-contextlib2==0.6.0.post1
-cryptography==2.1.4
-decorator==4.4.2
-defusedxml==0.6.0
-entrypoints==0.3
-enum34==1.1.9
-funcsigs==1.0.2
-functools32==3.2.3.post2
-futures==3.3.0
 idna==2.9
-importlib-metadata==1.5.0
-ipaddress==1.0.23
-ipykernel==4.10.1
-ipython==5.9.0
-ipython-genutils==0.2.0
-Jinja2==2.11.1
-jsonschema==3.2.0
-jupyter-client==5.3.4
-jupyter-core==4.6.3
-jupyterlab==0.33.12
-jupyterlab-launcher==0.11.2
-keyring==10.6.0
-keyrings.alt==3.0
-MarkupSafe==1.1.1
-mistune==0.8.4
-more-itertools==5.0.0
-nbconvert==5.6.1
-nbformat==4.4.0
-notebook==5.7.8
-olefile==0.45.1
-packaging==20.3
-pandocfilters==1.4.2
-pathlib2==2.3.5
-pathspec==0.7.0
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==5.1.0
-pluggy==0.13.1
-prometheus-client==0.7.1
-prompt-toolkit==1.0.18
-ptyprocess==0.6.0
-py==1.8.1
-pycrypto==2.6.1
-Pygments==2.5.2
-pygobject==3.26.1
-pyparsing==2.4.6
-pyrsistent==0.15.7
 pyTelegramBotAPI==3.6.7
-pytest==4.6.9
-python-dateutil==2.8.1
 python-dotenv==0.12.0
-pyxdg==0.25
-PyYAML==5.3
-pyzmq==19.0.0
 requests==2.23.0
-scandir==1.10.0
-SecretStorage==2.3.1
-Send2Trash==1.5.0
-simplegeneric==0.8.1
-singledispatch==3.4.0.3
 six==1.14.0
-subprocess32==3.5.4
-terminado==0.8.3
-testpath==0.4.4
-tornado==5.1.1
-traitlets==4.3.3
-typing==3.7.4.1
 urllib3==1.25.8
-urwid==2.0.1
-wcwidth==0.1.8
-webencodings==0.5.1
-yamllint==1.20.0
-zipp==1.2.0
